Updated CSVGenerator to not crash if file doesn't exist

+++ CSVGenerator.py
@@ -122,7 +122,8 @@ class CSVGenerator:
         self.stitchLocalCsvs(yearsToScrape = years_to_generate)
         filepath = f"data/training_data/training_data_{years_to_generate[0]}-{years_to_generate[-1]}.csv"
         outputFileName = f"data/training_features/training_features_{years_to_generate[0]}-{years_to_generate[-1]}.csv"
-        os.remove(outputFileName)
+        if Path(outputFileName).is_file():
+            os.remove(outputFileName)
         games_frame = pd.read_csv(filepath)
         for year in years_to_generate:
             self._year_to_generate = year
